+++ Assets/Animation/UI/LostSouls_Closed.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d699c4e060a054a4d81ae3782165a7e5
+guid: 355a984956d8f1b4a82a0f314305a4b8
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 7400000
