+++ Assets/Animation/UI/LostSouls_Open.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 023ff4824c752284e846ea3df2503a8b
+guid: 992f29666d325494bad7b11936a72ef7
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 7400000
