+++ Assets/Scripts/PlayerScripts/LostSoulManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: da110c87864be82458ea7793cf2270de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
